requirements.txt
secretsharing
pycrypto==2.6.1
git+https://github.com/blockstack/secret-sharing.git
